change procedural resource API, fix examples

+++ sources/Lunacub/ResourceID.cs
@@ -11,38 +11,38 @@ namespace Caxivitual.Lunacub;
 [DebuggerDisplay("{Value}")]
 public readonly struct ResourceID :
     IEquatable<ResourceID>, 
-    IEquatable<ulong>,
+    IEquatable<UInt128>,
     ISpanFormattable,
     IUtf8SpanFormattable,
     ISpanParsable<ResourceID>,
     IUtf8SpanParsable<ResourceID>
 {
     /// <summary>
-    /// The underlying 64-bit unsigned integer value of <see cref="ResourceID"/>.
+    /// The underlying 128-bit unsigned integer value of <see cref="ResourceID"/>.
     /// </summary>
-    [DebuggerBrowsable(DebuggerBrowsableState.Never)] public ulong Value { get; }
+    [DebuggerBrowsable(DebuggerBrowsableState.Never)] public UInt128 Value { get; }
     
     /// <summary>
-    /// Creates a new instance of <see cref="ResourceID"/> with the specified 64-bit unsigned integer value.
+    /// Creates a new instance of <see cref="ResourceID"/> with the specified 128-bit unsigned integer value.
     /// </summary>
-    /// <param name="value">The specified 64-bit unsigned integer value to create <see cref="ResourceID"/> from.</param>
-    public ResourceID(ulong value) {
+    /// <param name="value">The specified 128-bit unsigned integer value to create <see cref="ResourceID"/> from.</param>
+    public ResourceID(UInt128 value) {
         Value = value;
     }
 
     /// <summary>
     /// Parses a string into a <see cref="ResourceID"/> with formatting information object.
     /// </summary>
-    /// <param name="s">A string represents a 64-bit unsigned integer to parse.</param>
+    /// <param name="s">A string represents a 128-bit unsigned integer to parse.</param>
     /// <param name="formatProvider">An object contains formatting information about <paramref name="s"/>.</param>
     /// <returns>The result of parsing <paramref name="s"/>.</returns>
     public static ResourceID Parse(string s, IFormatProvider? formatProvider) {
-        return new(ulong.Parse(s, formatProvider));
+        return new(UInt128.Parse(s, formatProvider));
     }
     
     /// <inheritdoc cref="Parse(ReadOnlySpan{char},IFormatProvider?)"/>
     public static ResourceID Parse(ReadOnlySpan<byte> span, IFormatProvider? formatProvider) {
-        return new(ulong.Parse(span, formatProvider));
+        return new(UInt128.Parse(span, formatProvider));
     }
 
     /// <inheritdoc cref="Parse(ReadOnlySpan{char},NumberStyles,IFormatProvider?)"/>
@@ -51,23 +51,23 @@ public readonly struct ResourceID :
         NumberStyles style = NumberStyles.Integer,
         IFormatProvider? formatProvider = null
     ) {
-        return new(ulong.Parse(span, style, formatProvider));
+        return new(UInt128.Parse(span, style, formatProvider));
     }
     
     /// <summary>
     /// Parses a span of characters into a <see cref="ResourceID"/> with formatting information object.
     /// </summary>
-    /// <param name="span">A span of character represents a 64-bit unsigned integer to parse.</param>
+    /// <param name="span">A span of character represents a 128-bit unsigned integer to parse.</param>
     /// <param name="formatProvider">An object contains formatting information about <paramref name="span"/>.</param>
     /// <returns>The result of parsing <paramref name="span"/>.</returns>
     public static ResourceID Parse(ReadOnlySpan<char> span, IFormatProvider? formatProvider) {
-        return new(ulong.Parse(span, formatProvider));
+        return new(UInt128.Parse(span, formatProvider));
     }
 
     /// <summary>
     /// Parses a span of characters into a <see cref="ResourceID"/> with formatting information object.
     /// </summary>
-    /// <param name="span">A span of character represents a 64-bit unsigned integer to parse.</param>
+    /// <param name="span">A span of character represents a 128-bit unsigned integer to parse.</param>
     /// <param name="style">A bitwise combination of number styles that can be present in <paramref name="span"/>.</param>
     /// <param name="formatProvider">An object contains formatting information about <paramref name="span"/>.</param>
     /// <returns>The result of parsing <paramref name="span"/>.</returns>
@@ -76,12 +76,12 @@ public readonly struct ResourceID :
         NumberStyles style = NumberStyles.Integer,
         IFormatProvider? formatProvider = null
     ) {
-        return new(ulong.Parse(span, style, formatProvider));
+        return new(UInt128.Parse(span, style, formatProvider));
     }
     
     /// <inheritdoc cref="TryParse(ReadOnlySpan{char},IFormatProvider?,out Caxivitual.Lunacub.ResourceID)"/>
     public static bool TryParse(ReadOnlySpan<byte> span, IFormatProvider? formatProvider, out ResourceID result) {
-        if (ulong.TryParse(span, formatProvider, out var value)) {
+        if (UInt128.TryParse(span, formatProvider, out var value)) {
             result = new(value);
             return true;
         }
@@ -97,7 +97,7 @@ public readonly struct ResourceID :
         IFormatProvider? formatProvider,
         out ResourceID result
     ) {
-        if (ulong.TryParse(span, style, formatProvider, out var value)) {
+        if (UInt128.TryParse(span, style, formatProvider, out var value)) {
             result = new(value);
             return true;
         }
@@ -109,7 +109,7 @@ public readonly struct ResourceID :
     /// <summary>
     /// Tries to parse a span of character into a <see cref="ResourceID"/>.
     /// </summary>
-    /// <param name="span">A span of character represents a 64-bit unsigned integer to parse.</param>
+    /// <param name="span">A span of character represents a 128-bit unsigned integer to parse.</param>
     /// <param name="formatProvider">An object contains formatting information about <paramref name="span"/>.</param>
     /// <param name="result">
     ///     When this method returns, contains the result of successfully parsing <paramref name="span"/>, or the default
@@ -119,7 +119,7 @@ public readonly struct ResourceID :
     ///     <see langword="true"/> if <paramref name="span"/> was parsed successfully; otherwise, <see langword="false"/>.
     /// </returns>
     public static bool TryParse(ReadOnlySpan<char> span, IFormatProvider? formatProvider, out ResourceID result) {
-        if (ulong.TryParse(span, formatProvider, out var value)) {
+        if (UInt128.TryParse(span, formatProvider, out var value)) {
             result = new(value);
             return true;
         }
@@ -131,7 +131,7 @@ public readonly struct ResourceID :
     /// <summary>
     /// Tries to parse a span of character into a <see cref="ResourceID"/>.
     /// </summary>
-    /// <param name="span">A span of character represents a 64-bit unsigned integer to parse.</param>
+    /// <param name="span">A span of character represents a 128-bit unsigned integer to parse.</param>
     /// <param name="style">A bitwise combination of number styles that can be present in <paramref name="span"/>.</param>
     /// <param name="formatProvider">An object contains formatting information about <paramref name="span"/>.</param>
     /// <param name="result">
@@ -142,7 +142,7 @@ public readonly struct ResourceID :
     ///     <see langword="true"/> if <paramref name="span"/> was parsed successfully; otherwise, <see langword="false"/>.
     /// </returns>
     public static bool TryParse(ReadOnlySpan<char> span, NumberStyles style, IFormatProvider? formatProvider, out ResourceID result) {
-        if (ulong.TryParse(span, style, formatProvider, out var value)) {
+        if (UInt128.TryParse(span, style, formatProvider, out var value)) {
             result = new(value);
             return true;
         }
@@ -154,7 +154,7 @@ public readonly struct ResourceID :
     /// <summary>
     /// Tries to parse a string into a <see cref="ResourceID"/>.
     /// </summary>
-    /// <param name="s">A string represents a 64-bit unsigned integer to parse.</param>
+    /// <param name="s">A string represents a 128-bit unsigned integer to parse.</param>
     /// <param name="formatProvider">An object contains formatting information about <paramref name="s"/>.</param>
     /// <param name="result">
     ///     When this method returns, contains the result of successfully parsing <paramref name="s"/>, or the default
@@ -170,7 +170,7 @@ public readonly struct ResourceID :
     /// <summary>
     /// Tries to parse a string into a <see cref="ResourceID"/>.
     /// </summary>
-    /// <param name="s">A string represents a 64-bit unsigned integer to parse.</param>
+    /// <param name="s">A string represents a 128-bit unsigned integer to parse.</param>
     /// <param name="style">A bitwise combination of number styles that can be present in <paramref name="s"/>.</param>
     /// <param name="formatProvider">An object contains formatting information about <paramref name="s"/>.</param>
     /// <param name="result">
@@ -257,32 +257,32 @@ public readonly struct ResourceID :
     public bool Equals(ResourceID other) => this == other;
 
     /// <summary>
-    /// Determines whether the <see cref="Value"/> of the current instance is equal to the specified 64-bit unsigned
+    /// Determines whether the <see cref="Value"/> of the current instance is equal to the specified 128-bit unsigned
     /// integer.
     /// </summary>
-    /// <param name="other">The 64-bit unsigned integer to compare with the <see cref="Value"/> of the current instance.</param>
+    /// <param name="other">The 128-bit unsigned integer to compare with the <see cref="Value"/> of the current instance.</param>
     /// <returns>
-    ///     <see langword="true"/> if the <see cref="Value"/> of the current instance is equal to the provided 64-bit
+    ///     <see langword="true"/> if the <see cref="Value"/> of the current instance is equal to the provided 128-bit
     ///     unsigned integer; otherwise, <see langword="false"/>
     /// </returns>
-    public bool Equals(ulong other) => this == other;
+    public bool Equals(UInt128 other) => this == other;
 
     /// <summary>
     /// Determines whether this instance and a specified object, which can be either an instance of <see cref="ResourceID"/>
-    /// or a 64-bit unsigned integer, are equal.
+    /// or a 128-bit unsigned integer, are equal.
     /// </summary>
     /// <param name="other">The object to compare to.</param>
     /// <returns>
     ///     <see langword="true"/> if <paramref name="other"/> is a <see cref="ResourceID"/> and is equal to the current instance,
-    ///     or if <paramref name="other"/> is a 64-bit unsigned integer and is equal to the <see cref="Value"/> of the current
+    ///     or if <paramref name="other"/> is a 128-bit unsigned integer and is equal to the <see cref="Value"/> of the current
     ///     instance; otherwise, <see langword="false"/>.
     /// </returns>
     /// <seealso cref="Equals(ResourceID)"/>
-    /// <seealso cref="Equals(ulong)"/>
+    /// <seealso cref="Equals(UInt128)"/>
     public override bool Equals([NotNullWhen(true)] object? other) {
         return other switch {
             ResourceID rid => Equals(rid),
-            ulong u128 => Equals(u128),
+            UInt128 u128 => Equals(u128),
             _ => false,
         };
     }
@@ -291,12 +291,12 @@ public readonly struct ResourceID :
 
     public static bool operator ==(ResourceID left, ResourceID right) => left.Value == right.Value;
     public static bool operator !=(ResourceID left, ResourceID right) => left.Value != right.Value;
-    public static bool operator ==(ResourceID left, ulong right) => left.Value == right;
-    public static bool operator !=(ResourceID left, ulong right) => left.Value != right;
+    public static bool operator ==(ResourceID left, UInt128 right) => left.Value == right;
+    public static bool operator !=(ResourceID left, UInt128 right) => left.Value != right;
     
     public static implicit operator ResourceID(uint value) => new(value);
-    public static implicit operator ResourceID(ulong value) => new(value);
-    public static implicit operator ulong(ResourceID value) => value.Value;
+    public static implicit operator ResourceID(UInt128 value) => new(value);
+    public static implicit operator UInt128(ResourceID value) => value.Value;
 
     /// <summary>
     /// Returns the string representation of this instance in default decimal integer format.
